// frontend/src/app/models/zahtev.ts
export class Zahtev{
  idZ: number= 0;
  kor_ime: string = '';
  lozinka: string = '';
  ime: string = '';
  prezime: string = '';
  mejl: string = '';
  tip: string = '';
  pol: string = '';
  adresa: string = '';
  kontakt: string = '';
  //brKartice: string = '';
  profilna: string = '';
  status: string = '';
}